pyproject.toml
[project]
name = "fed-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
